requirements.txt
# Core dependencies
crewai>=0.28.0
langchain>=0.0.267
langchain-openai>=0.0.2
langchain-anthropic>=0.1.1
langchain-mistralai>=0.0.1
langchain-community>=0.0.10

# API integrations
openai>=1.3.0
anthropic>=0.5.0

# Research tools dependencies
yfinance>=0.2.31
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2

# Environment and utilities
python-dotenv>=1.0.0

# Optional search tools
duckduckgo-search>=3.9.3
brave-search>=0.1.2
firecrawl-py 'crewai[tools]'
nltk
gnews